Move actions pkg at server root
